# internal/api/wellknown/openapi.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "See docs/10-running-in-production.md for token issuance. A route with no matching scope in the token, and no admin role, is refused with 403.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Every route declared in internal/apispec, the package cmd/controller/main.go also builds its real router from, served under /api/v1. A listed route's method, pattern, scope, and link relation are the ones the server enforces. That a listed route is registered at all is not checked by anything, so confirm a route against a running controller before building on it.",
    "title": "Pleiades control plane API",
    "version": "unreleased"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/bindings": {
      "get": {
        "description": "Returns bindings, keyset-paged. Each is one grant: a team, a role, a place in the containment hierarchy, and allow or deny.",
        "operationId": "list_bindings",
        "parameters": [
          {
            "description": "Keyset cursor: the last id of the previous page.",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Maximum records to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Case-insensitive name filter.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "_links": {
                      "description": "Hypermedia affordances available on this resource for the calling identity.",
                      "items": {
                        "properties": {
                          "href": {
                            "type": "string"
                          },
                          "method": {
                            "type": "string"
                          },
                          "rel": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "bindings": {
                      "items": {
                        "properties": {
                          "_links": {
                            "description": "Hypermedia affordances available on this resource for the calling identity.",
                            "items": {
                              "properties": {
                                "href": {
                                  "type": "string"
                                },
                                "method": {
                                  "type": "string"
                                },
                                "rel": {
                                  "type": "string"
                                }
                              },
                              "type": "object"
                            },
                            "type": "array"
                          },
                          "effect": {
                            "description": "allow or deny. An explicit deny at any level beats an allow at a broader one.",
                            "type": "string"
                          },
                          "granted_at": {
                            "description": "Where this grant sits, rendered for a reader: the scope type and the named record.",
                            "type": "string"
                          },
                          "id": {
                            "description": "Stable numeric identifier.",
                            "type": "integer"
                          },
                          "role": {
                            "description": "viewer, operator or admin.",
                            "type": "string"
                          },
                          "scope_id": {
                            "description": "The record this grant covers. Absent for system scope, which names no target.",
                            "type": "integer"
                          },
                          "scope_name": {
                            "description": "The named record this grant covers. Absent for system scope, which names no target, and for a target that has been deleted.",
                            "type": "string"
                          },
                          "scope_type": {
                            "description": "system, organization, inventory, group or device.",
                            "type": "string"
                          },
                          "team": {
                            "description": "The team this grant is held by. Roles are never granted to a user directly.",
                            "type": "integer"
                          },
                          "team_name": {
                            "description": "That team's name, beside its id, so a client can render the reference without a second request.",
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "A page of bindings."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No identity on the request context."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The caller lacks access:read."
          }
        },
        "security": [
          {
            "bearerAuth": [
              "access:read"
            ]
          }
        ],
        "summary": "List bindings",
        "x-link-relation": "collection"
      },
      "post": {
        "description": "Creates one grant: a team, a role, a place in the containment hierarchy, and allow or deny.",
        "operationId": "create_role_binding",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "_links": {
                      "description": "Hypermedia affordances available on this resource for the calling identity.",
                      "items": {
                        "properties": {
                          "href": {
                            "type": "string"
                          },
                          "method": {
                            "type": "string"
                          },
                          "rel": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "effect": {
                      "description": "allow or deny. An explicit deny at any level beats an allow at a broader one.",
                      "type": "string"
                    },
                    "granted_at": {
                      "description": "Where this grant sits, rendered for a reader: the scope type and the named record.",
                      "type": "string"
                    },
                    "id": {
                      "description": "Stable numeric identifier.",
                      "type": "integer"
                    },
                    "role": {
                      "description": "viewer, operator or admin.",
                      "type": "string"
                    },
                    "scope_id": {
                      "description": "The record this grant covers. Absent for system scope, which names no target.",
                      "type": "integer"
                    },
                    "scope_name": {
                      "description": "The named record this grant covers. Absent for system scope, which names no target, and for a target that has been deleted.",
                      "type": "string"
                    },
                    "scope_type": {
                      "description": "system, organization, inventory, group or device.",
                      "type": "string"
                    },
                    "team": {
                      "description": "The team this grant is held by. Roles are never granted to a user directly.",
                      "type": "integer"
                    },
                    "team_name": {
                      "description": "That team's name, beside its id, so a client can render the reference without a second request.",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The created record."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The submission is not a valid role binding."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "That name is already taken."
          }
        },
        "security": [
          {
            "bearerAuth": [
              "access:write"
            ]
          }
        ],
        "summary": "Create a role binding",
        "x-link-relation": "create"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Relative to the controller's own origin.",
      "url": "/api/v1"
    }
  ]
}